feat(fe): remove legacy fragment callback page
The IdP's form_post response is answered by the canister directly, so
the /callback route never renders: its GET page, the sendUrlToOpener
helper and the layout's origin-redirect exclusion for the path are dead.
The live popup utilities (redirectInPopup, CallbackPopupClosedError,
CallbackPayload) move from the route folder into $lib/utils/openID.ts
next to their only consumers.
